feat: move to hasher/ module
