+++ hasher/catalogue-consts.h
