+++ hasher/core-types.cpp
@@ -1,4 +1,4 @@
-#include "core-types.h"
+#include "hasher/core-types.h"
 
 void slice_hex(const Slice sl, std::string &result) {
     static constexpr char hexChars[] = "0123456789abcdef";
