+++ hasher/core-types.h
