+++ hasher/log-macros.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "logger.h"
+#include "hasher/logger.h"
 
 // Macro for logging hashes efficiently (only formats if DEBUG is enabled)
 #define LOGD_KEY(label, key_obj)    if(Logger::getLevel() >= LogLevel::DEBUG) Logger::logWithFormat(LogLevel::DEBUG, \
