+++ hasher/logger.h
