+++ hasher/utils.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "hasher/utils.h"
 #include <ctime>
 
 namespace utils {
