+++ hasher/utils.h
